Split up testCMC* into multiple tests.
This allows us to run the tests substantially faster when multiple cores are
available. In particular, this will make the travis tests run quicker.
