+++ Applications/CMC/test/CMakeLists.txt
@@ -8,17 +8,38 @@ LINK_LIBRARIES( debug osimCommon${CMAKE_DEBUG_POSTFIX} optimized osimCommon
    		debug ${NameSpace}SimTKcommon${CMAKE_DEBUG_POSTFIX} optimized   ${NameSpace}SimTKcommon
    		debug ${NameSpace}SimTKmath${CMAKE_DEBUG_POSTFIX} optimized  ${NameSpace}SimTKmath)
 
-ADD_EXECUTABLE(testCMC testCMC.cpp)
-ADD_EXECUTABLE(testCMCWithControlConstraints testCMCWithControlConstraints.cpp)
+ADD_CUSTOM_TARGET(copyCMCTestFiles)
+
+IF (EXECUTABLE_OUTPUT_PATH)
+  SET (TEST_PATH ${EXECUTABLE_OUTPUT_PATH})
+ELSE (EXECUTABLE_OUTPUT_PATH)
+  SET (TEST_PATH .)
+ENDIF (EXECUTABLE_OUTPUT_PATH)
+
+FILE(GLOB TEST_PROGS "test*.cpp")
+
+FOREACH(TEST_PROG ${TEST_PROGS})
+
+    GET_FILENAME_COMPONENT(TEST_ROOT ${TEST_PROG} NAME_WE)
+
+    ADD_EXECUTABLE(${TEST_ROOT} ${TEST_PROG})
+    TARGET_LINK_LIBRARIES(${TEST_ROOT} ${LINK_LIBRARIES})
+    ADD_TEST(${TEST_ROOT} ${TEST_PATH}/${TEST_ROOT})
+    SET_TARGET_PROPERTIES(${TEST_ROOT}
+        PROPERTIES ${EXCLUDE_IF_MINIMAL_BUILD}
+        PROJECT_LABEL "Tests - ${TEST_ROOT}")
+
+ENDFOREACH()
 
 FILE(GLOB TEST_FILES *.osim *.xml *.sto *.mot)
 
 #
 # Copy files to run dir
 #
+ADD_DEPENDENCIES(${TEST_ROOT} copyCMCTestFiles)
 FOREACH (dataFile ${TEST_FILES})
  ADD_CUSTOM_COMMAND(
-    TARGET testCMC
+    TARGET copyCMCTestFiles
     COMMAND ${CMAKE_COMMAND}
     ARGS -E copy
     ${dataFile}
@@ -35,16 +56,3 @@ FOREACH (dataFile ${TEST_FILES})
 #ENDIF (UNIX)
 
 #ENABLE_TESTING()
-
-IF (EXECUTABLE_OUTPUT_PATH)
-  SET (TEST_PATH ${EXECUTABLE_OUTPUT_PATH})
-ELSE (EXECUTABLE_OUTPUT_PATH)
-  SET (TEST_PATH .)
-ENDIF (EXECUTABLE_OUTPUT_PATH)
-
-
-ADD_TEST(testCMC ${TEST_PATH}/testCMC)
-ADD_TEST(testCMCWithControlConstraints ${TEST_PATH}/testCMCWithControlConstraints)
-
-SET_TARGET_PROPERTIES(testCMC PROPERTIES ${EXCLUDE_IF_MINIMAL_BUILD} PROJECT_LABEL "Tests - testCMC")
-SET_TARGET_PROPERTIES(testCMCWithControlConstraints PROPERTIES ${EXCLUDE_IF_MINIMAL_BUILD} PROJECT_LABEL "Tests - testCMCWithControlConstraints")
